fix #43 연관관계 mapping Community && DailyMeal ver. 08 조회시 수정시간 변하는 부분 수정

+++ main_project/src/main/java/NutrientsCoders/main_project/community/entity/Community.java
@@ -28,7 +28,7 @@ public class Community extends CommunityBaseTime {
     @Column
     private long recommendationCount = 0L;
     @Column
-    private long communityViewCount = 0L;
+    private int communityViewCount = 0;
     @Column
     private int communityLike = 0;
     @Column
@@ -37,6 +37,8 @@ public class Community extends CommunityBaseTime {
     @CollectionTable(name = "MEMBER_ID", joinColumns = @JoinColumn(name = "member_Id"))
     @Column(name = "LIKE_MEMBERS",insertable = false)
     private List<Long> likeMembers;
+
+
     @JsonIgnore
     @ManyToOne
     @JoinColumn(name = "DAILYMEAL_ID", nullable = true)
